feat: 쪼개져있던 두 Search페이지 통합
